feature: Unifies the engine↔TUI formatter seam
