+++ packages/engine/src/fund-composition.ts
@@ -533,77 +533,6 @@ export function buildDashboardDetail(
   };
 }
 
-export function formatCompositionReport(report: CompositionReport): string {
-  const sections = [
-    `Numisma Fund Composition Prototype`,
-    divider(),
-    `Fund: ${report.dashboard.summary.fundName}`,
-    `As of: ${report.dashboard.summary.asOf}`,
-    "",
-    formatWeeklyReviewFocus(report),
-    "",
-    "Canonical Summary",
-    `Fund value: ${formatUsd(report.totals.fundValueUsd)}`,
-    `Unrealized P&L: ${formatUsd(report.dashboard.summary.totalUnrealizedPnlUsd)}`,
-    `Manual FX: 1 USD = ${report.totals.usdMxn.toFixed(4)} MXN`,
-    `Mode filter: live only; ${report.excluded.nonLive} non-live record(s) excluded`,
-    `Record safety: ${report.excluded.invalid} invalid record(s) excluded`,
-    ...(report.excluded.shortDeferred > 0
-      ? [
-          `Short direction: ${report.excluded.shortDeferred} deferred short record(s) excluded`,
-        ]
-      : []),
-    "",
-    formatRows("Portfolio Composition", sectionRows(report, "portfolios")),
-    "",
-    formatRows("Tempo Composition", sectionRows(report, "tempos")),
-    "",
-    formatRows("Account Composition", sectionRows(report, "accounts")),
-    "",
-    formatRows(
-      "Instrument Composition",
-      sectionRows(report, "instruments"),
-      true,
-    ),
-    "",
-    formatRows(
-      "Capital Tier Composition",
-      sectionRows(report, "tiers"),
-      true,
-    ),
-    "",
-    formatPriceJourneys(report.priceJourneys),
-  ];
-
-  if (report.warnings.length > 0) {
-    sections.push(
-      "",
-      "!!! WARNINGS !!!",
-      ...report.warnings.map((warning) => `- ${warning.message}`),
-    );
-  }
-
-  sections.push("", "Keys: q quit | r reload data");
-  return sections.join("\n");
-}
-
-function formatWeeklyReviewFocus(report: CompositionReport): string {
-  const { summary } = report.dashboard;
-
-  return [
-    "Weekly Review Focus",
-    "-------------------",
-    `Fund now: ${formatUsd(summary.fundValueUsd)} in live canonical records`,
-    `Unrealized P&L: ${formatUsd(summary.totalUnrealizedPnlUsd)}`,
-    `Largest Portfolio: ${formatRowFocus(summary.largestPortfolio)}`,
-    `Largest Tempo: ${formatRowFocus(summary.largestTempo)}`,
-    `Largest Account: ${formatRowFocus(summary.largestAccount)}`,
-    `Largest Instrument: ${formatRowFocus(summary.largestInstrument)}`,
-    `Reserve: ${formatRowFocus(summary.reserve)}`,
-    `Data safety: ${formatDataSafety(report)}`,
-  ].join("\n");
-}
-
 function buildCanonicalState(data: FundReviewData): CanonicalState {
   const warnings: Warning[] = [];
   const portfolios = indexById(data.portfolios, "portfolio");
@@ -923,33 +852,6 @@ function detailLinesForRow(
   return accountLines.filter((line) => line.recordKind === "position");
 }
 
-function formatRowFocus(row: DashboardFocus | undefined): string {
-  if (!row) return "No live records";
-  return `${row.label} ${formatUsd(row.usdValue)} (${formatPercent(row.percentOfFund)})`;
-}
-
-function formatDataSafety(report: CompositionReport): string {
-  const exclusions = [
-    `${report.excluded.nonLive} non-live excluded`,
-    `${report.excluded.invalid} invalid excluded`,
-    ...(report.excluded.shortDeferred > 0
-      ? [`${report.excluded.shortDeferred} short deferred`]
-      : []),
-  ];
-  return report.warnings.length > 0
-    ? `${exclusions.join("; ")}; warnings shown below`
-    : `${exclusions.join("; ")}; no warnings`;
-}
-
-function sectionRows(
-  report: CompositionReport,
-  sectionId: DashboardSectionId,
-): CompositionRow[] {
-  return (
-    report.dashboard.sections.find((section) => section.id === sectionId)?.rows ?? []
-  );
-}
-
 function toFocus(row: CompositionRow | undefined): DashboardFocus | undefined {
   if (!row) return undefined;
   return {
@@ -1122,86 +1024,6 @@ function groupTierLines(
       (a, b) =>
         Math.abs(b.usdValue) - Math.abs(a.usdValue) || a.label.localeCompare(b.label),
     );
-}
-
-function formatRows(
-  title: string,
-  rows: CompositionRow[],
-  includePnl = false,
-): string {
-  const header = includePnl
-    ? `${pad("Name", 28)} ${padLeft("USD Value", 14)} ${padLeft("Fund %", 8)} ${padLeft("Cost", 14)} ${padLeft("Unrl P&L", 14)}`
-    : `${pad("Name", 28)} ${padLeft("USD Value", 14)} ${padLeft("Fund %", 8)}`;
-  const body = rows.map((row) => {
-    const base = `${pad(row.label, 28)} ${padLeft(formatUsd(row.usdValue), 14)} ${padLeft(formatPercent(row.percentOfFund), 8)}`;
-    if (!includePnl) return base;
-    return `${base} ${padLeft(formatMaybeUsd(row.costBasisUsd), 14)} ${padLeft(formatMaybeUsd(row.unrealizedPnlUsd), 14)}`;
-  });
-
-  return [
-    title,
-    "-".repeat(title.length),
-    header,
-    "-".repeat(header.length),
-    ...(body.length > 0 ? body : ["No live records."]),
-  ].join("\n");
-}
-
-function formatPriceJourneys(journeys: PriceJourney[]): string {
-  const title = "Weekly Price Journey";
-  if (journeys.length === 0) {
-    return [title, "-".repeat(title.length), "No Close history recorded."].join("\n");
-  }
-
-  const body = journeys.map((journey) => {
-    const series = journey.points
-      .map((point) => `${point.asOf} ${formatPrice(point.price, journey.currency)}`)
-      .join(" -> ");
-    return `${pad(journey.label, 28)} ${series}  (${formatSignedPercent(journey.changePct)})`;
-  });
-
-  return [title, "-".repeat(title.length), ...body].join("\n");
-}
-
-function formatPrice(value: number, currency: Currency): string {
-  return new Intl.NumberFormat("en-US", {
-    style: "currency",
-    currency,
-  }).format(value);
-}
-
-function formatSignedPercent(value: number): string {
-  const sign = value > 0 ? "+" : "";
-  return `${sign}${value.toFixed(1)}%`;
-}
-
-function divider(): string {
-  return "=".repeat(36);
-}
-
-function formatUsd(value: number): string {
-  return new Intl.NumberFormat("en-US", {
-    style: "currency",
-    currency: "USD",
-  }).format(value);
-}
-
-function formatMaybeUsd(value: number | undefined): string {
-  return value === undefined ? "-" : formatUsd(value);
-}
-
-function formatPercent(value: number): string {
-  return `${value.toFixed(1)}%`;
-}
-
-function pad(value: string, width: number): string {
-  return value.length > width
-    ? `${value.slice(0, width - 1)}~`
-    : value.padEnd(width);
-}
-
-function padLeft(value: string, width: number): string {
-  return value.length > width ? value.slice(0, width) : value.padStart(width);
 }
 
 function accountLabel(
