+++ packages/engine/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./fund-composition.js";
+export * from "./format.js";
